renamed from AreaChart to Area
